Renames to match repo naming conventions
